% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathy}
\alias{bathy}
\title{Southern Newfoundland bathymetry}
\format{
A stars object

Derived from data downloaded from http://www.gebco.net/. Details provided in
the data-raw folder for this package.
}
\usage{
bathy
}
\description{
Southern Newfoundland bathymetry
}
\keyword{datasets}
