% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{convert_N}
\alias{convert_N}
\title{Convert abundance-at-age matrix to abundance-at-length}
\usage{
convert_N(N_at_age = NULL, lak = NULL)
}
\arguments{
\item{lak}{Length-age-key (i.e. probability of being in a specific length group given age)}

\item{N}{Abundance-at-age matrix}
}
\value{
Returns abundance-at-length matrix
}
\description{
Function for converting abundance-at-age matrix to abundance-at-length given
a length-age-key. Expects matrices to be named.
}
