% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{error_stats}
\alias{error_stats}
\title{Calculate common error statistics}
\usage{
error_stats(error)
}
\arguments{
\item{error}{A numeric vector of errors.}
}
\value{
A named vector of error statistics including:
\itemize{
\item \strong{ME}: Mean error
\item \strong{MAE}: Mean absolute error
\item \strong{MSE}: Mean squared error
\item \strong{RMSE}: Root mean squared error
}
}
\description{
Calculate common error statistics
}
