% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{gaussian_fun}
\alias{gaussian_fun}
\title{Define relationships with covariates}
\usage{
gaussian_fun(alpha = 1, mu = 250, sigma = 100, plot = FALSE)
}
\arguments{
\item{alpha, mu, sigma}{Parameters}

\item{plot}{Produce a simple plot of the simulated values?}
}
\description{
Closure to be used in \code{\link{sim_distribution}}
}
