% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{group_lengths}
\alias{group_lengths}
\title{Convert length to length group}
\usage{
group_lengths(length, group)
}
\arguments{
\item{length}{Numeric vector of lengths to be grouped. Used with \code{\link[base:findInterval]{base::findInterval()}}.}

\item{group}{Numeric value specifying the width of the length group (i.e., bin size).}
}
\value{
A numeric vector indicating the midpoint of the assigned length group.
}
\description{
Helper function for converting lengths to length groups.
\strong{Note:} This is not a general-purpose function — the output midpoints defining the groups
are aligned with DFO-specific methods and labeling conventions.
}
