% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculate intraclass correlation}
\usage{
icc(x, group)
}
\arguments{
\item{x}{Response variable.}

\item{group}{Grouping variable.}
}
\value{
An estimate of intraclass correlation.
}
\description{
A simple function for calculating intraclass correlation using \code{\link[lme4:lmer]{lme4::lmer()}}.
The formula follows the description provided on \href{https://en.wikipedia.org/wiki/Intraclass_correlation}{Wikipedia}.
}
