% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Southern Newfoundland coastline}
\format{
A sf object (MULTIPOLYGON)

Derived from global administrative boundaries data (http://gadm.org/), which
was downloaded using the now deprecated function, \code{raster::getData()}. Details provided in the
data-raw folder for this package.
}
\usage{
land
}
\description{
Southern Newfoundland coastline
}
\keyword{datasets}
