% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{linear_fun}
\alias{linear_fun}
\alias{parabolic_fun}
\title{Define relationships with covariates}
\usage{
linear_fun(alpha = 0, beta = NULL, scale = TRUE)

parabolic_fun(mu = 200, sigma = 100, scale = TRUE)
}
\arguments{
\item{beta, mu, sigma}{Parameters}

\item{scale}{Center effect around zero?}
}
\description{
Simple closures used to define relationships with covariates
}
