% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strap.R
\name{mid_length}
\alias{mid_length}
\title{Convert cut labels to mid points}
\usage{
mid_length(cut_labs, length_group)
}
\arguments{
\item{cut_labs}{Labels from \code{\link{base::cut}}}

\item{length_group}{Length group used to cut the length data}
}
\description{
Helper function for converting \code{\link{base::cut}} labels
into length mid points provided length group (Note: this isn't a general function;
output alligns with DFO specific method/labeling)
}
