% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{object_size}
\alias{object_size}
\title{Print object size}
\usage{
object_size(x, units = "Mb")
}
\arguments{
\item{x}{an \code{R} object}

\item{units}{the units to be used in printing the size}
}
\description{
A wrapper for \code{\link[utils]{object.size}} that prints in Mb
by default
}
