% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{parabola_fun}
\alias{parabola_fun}
\title{Define relationships with covariates}
\usage{
parabola_fun(alpha = 0, mu = 250, sigma = 50, plot = FALSE)
}
\arguments{
\item{alpha, mu, sigma}{Parameters that control the shape of the parabola}

\item{plot}{Produce a simple plot of the simulated values?}
}
\description{
Closure to be used in \code{\link{sim_distribution}}
}
