% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_surveys.R
\name{resume_test}
\alias{resume_test}
\title{Resume partial run of \code{\link{test_surveys}}}
\usage{
resume_test(export_dir = NULL)
}
\arguments{
\item{export_dir}{Export directory specified when \code{\link{test_surveys}} was run.}
}
\description{
Resume partial run of \code{\link{test_surveys}}
}
\details{
Progress bar time estimates will be biased here by previous completions
}
