% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_covar}
\alias{sim_covar}
\title{Simulate exponential or Matern covariance}
\usage{
sim_covar(range = NULL, variance = 1, model = "exponential")
}
\arguments{
\item{range}{Decorrelation range}

\item{variance}{Spatial variance}

\item{model}{String indicating either "exponential" or "matern" as the correlation function}
}
\description{
These functions return a function to use inside \code{\link{sim_distribution}}.
}

