% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_distribution}
\alias{sim_distribution}
\title{Simulate spatial and temporal distribution}
\usage{
sim_distribution(pop = sim_abundance(), grid = survey_grid,
  size_covar = sim_exp_covar(range = 4), time_covar = sim_exp_covar(range =
  4), space_covar = sim_exp_covar(range = 200, psill = 5))
}
\arguments{
\item{grid}{A \code{\link{SpatialPolygonsDataFrame}} defining a regular or irregular
grid with the same structure as \code{\link{survey_grid}}}

\item{N}{An abundance at age matrix with ages defining the rows and years defining
the columns (i.e. same structure as a matrix provided by \code{\link{sim_abundance}})}
}
\description{
Provided an abundance at age matrix (like one provided by \code{\link{sim_abundance}})
and a survey grid (like \code{\link{survey_grid}}) to populate, this function
applies correlated space, time and size error to simulate the distribution
of the population.
}
\examples{
sim_distribution()

}

