% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_exp_covar}
\alias{sim_exp_covar}
\title{Simulate exponential covariance}
\usage{
sim_exp_covar(range = NULL, psill = 1)
}
\arguments{
\item{range}{Decorrelation range (years, age or km)}

\item{psill}{Partial sill}
}
\description{
These function returns a function to use inside \code{\link{sim_distribution}}.
}

