% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_growth.R
\name{sim_growth}
\alias{sim_growth}
\title{Simulate length using von Bertalanffy notation}
\usage{
sim_growth(Linf = 120, L0 = 5, K = 0.1, log_sd = 0.1, plot = FALSE)
}
\arguments{
\item{Linf}{Mean asymptotic length}

\item{L0}{Length at birth}

\item{K}{Growth rate parameter}

\item{log_sd}{Standard deviation of the relationship in log scale}

\item{plot}{Produce a simple plot of the simulated values?}
}
\description{
This function returns a function to use inside \code{\link{sim_distribution}}.
Given parameters, it generates an approximate age-length key and inverse
age-length key by sampling from a von Bertalanffy relationship. Note that length is
discretized by rounding.
}
