% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_pop}
\alias{sim_pop}
\title{Simulate an age-structured population over a survey grid.}
\usage{
sim_pop(ages = 1:6, grid = survey_grid)
}
\arguments{
\item{ages}{Ages to include in the simulated population.}

\item{grid}{Survey grid to populate with simulated data.}
}
\value{
A 'data.frame' containing ages and lengths for every fish in the
    simulated population.
}
\description{
Simulate age-structured population that varies in space and time
}
\examples{
sim_pop()
}
