% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_sp_covar}
\alias{sim_sp_covar}
\alias{sim_ay_covar}
\title{Simulate exponential or Matern covariance}
\usage{
sim_sp_covar(range = 200, sd = 0.5, model = "matern")

sim_ay_covar(sd = 0.5, phi_age = 0.7, phi_year = 0.9)
}
\arguments{
\item{range}{Decorrelation range}

\item{sd}{Spatial variance}

\item{model}{String indicating either "exponential" or "matern" as the correlation function}
}
\description{
These functions return a function to use inside \code{\link{sim_distribution}}.
}
