% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_time_cor}
\alias{sim_size_cor}
\alias{sim_space_cor}
\alias{sim_time_cor}
\title{Simulate size, time and space correlation structure}
\usage{
sim_time_cor(dcor_time = 2)

sim_size_cor(dcor_size = 4)

sim_space_cor(tau = 0.25, theta = 0.001)
}
\arguments{
\item{dcor_time}{Decorrelation time (years)}

\item{dcor_size}{Decorrelation size (age)}

\item{tau, theta}{Precision matrix parameters}

\item{dcor_dist}{Decorrelation distance (km)}

\item{sigma}{Spatial variance}
}
\description{
These function returns a function to use inside \code{\link{sim_distribution}}.
Given parameters, it generates recruitment values from a log normal distribution
}
\details{
The formulation of these functions follow Kristensen et al. (2013).
}
\references{
Kristensen, K., Thygesen, U. H., Andersen, K. H., & Beyer, J. E.
(2013). Estimating spatio-temporal dynamics of size-structured populations.
Canadian Journal of Fisheries and Aquatic Sciences, 71(2), 326-336.
}

