% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPop.R
\name{SimAbundance}
\alias{SimAbundance}
\title{Simulate basic population dynamics model}
\usage{
SimAbundance(ages = 1:6, years = 1:10, Z = 0.2, r = 1000)
}
\arguments{
\item{ages}{Ages to include in the simulation.}

\item{years}{Years to include in the simulation.}

\item{Z}{Total mortality. Can be one value or a matrix of \code{length(ages)}
rows and \code{length(years)} columns.}

\item{r}{Recruitment (i.e. Abundance at \code{min(ages)}). Can be one value
or vector of \code{length(years)}}
}
\value{
A \code{matrix} of abundance at age by year.
}
\description{
Simulate basic population dynamics model
}
\details{
Abundance from \code{ages[2:max(ages)]} and \code{years[2:max(years)]} is
calculated using a standard population dynamics model:
\deqn{N_{a, y} = N_{a - 1, y - 1} * exp(-Z_{a - 1, y - 1})}{N_a,y = N_a-1,y-1 * exp(-Z_a-1,y-1)}
Abundance at \code{min(ages)} is supplied by r and abundance at \code{ages[2:max(ages)]} are
calculated using the same equation above using Z values from \code{min(years)}.
}
\examples{
SimAbundance()

}

