% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simDistribution}
\alias{simDistribution}
\title{Simulate spatial and temporal distribution}
\usage{
simDistribution(pop = simAbundance(), grid = survey_grid, rho = 0.05,
  sigma = 1e-05)
}
\arguments{
\item{grid}{A \code{\link{SpatialPolygonsDataFrame}} defining a regular or irregular
grid with the same structure as \code{\link{survey_grid}}}

\item{N}{An abundance at age matrix with ages defining the rows and years defining
the columns (i.e. same structure as a matrix provided by \code{\link{simAbundance}})}
}
\description{
Provided an abundance at age matrix (like one provided by \code{\link{simAbundance}})
and a survey grid (like \code{\link{survey_grid}}) to populate, this function
applies spatial and temporal error to simulate the spatial and temporal distribution
of the population.
}
\examples{
simDistribution()

}

