% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPop.R
\name{SimPop}
\alias{SimPop}
\title{Simulate an age-structured population over a survey grid.}
\usage{
SimPop(ages = 1:6, grid = survey_grid)
}
\arguments{
\item{ages}{Ages to include in the simulated population.}

\item{grid}{Survey grid to populate with simulated data.}
}
\value{
A 'data.frame' containing ages and lengths for every fish in the
    simulated population.
}
\description{
Simulate an age-structured population over a survey grid.
}
\examples{
simPop()
}

