% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simTime}
\alias{simSize}
\alias{simSpace}
\alias{simSpacePM}
\alias{simTime}
\title{Simulate size, time and space correlation structure}
\usage{
simTime(dcor_time = 2)

simSize(dcor_size = 4)

simSpace(dcor_dist = 70, sigma = 1.5)

simSpacePM(tau = 0.25, theta = 0.001)
}
\arguments{
\item{dcor_time}{Decorrelation time (years)}

\item{dcor_size}{Decorrelation size (age)}

\item{dcor_dist}{Decorrelation distance (km)}

\item{sigma}{Spatial variance}

\item{tau, theta}{Precision matrix parameters}
}
\description{
These function returns a function to use inside \code{\link{simDistribution}}.
Given parameters, it generates recruitment values from a log normal distribution
}
\details{
The formulation of these functions follow Kristensen et al. (2013).
}
\references{
Kristensen, K., Thygesen, U. H., Andersen, K. H., & Beyer, J. E.
(2013). Estimating spatio-temporal dynamics of size-structured populations.
Canadian Journal of Fisheries and Aquatic Sciences, 71(2), 326-336.
}

