% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simZ}
\alias{simZ}
\title{Simulate total mortality using a random walk}
\usage{
simZ(mean = 0.4, sd = 0.2, breaks = NULL)
}
\arguments{
\item{mean, sd}{Mean and standard deviation of total mortality.}

\item{breaks}{Provide breaks to age and/or year series to specify group specific
mean total mortality values. If specified, multiple mean values must be
supplied; one more than the number of breaks. Provide named lists (see
examples below).}
}
\description{
This function returns a function to use inside \code{\link{simAbundance}}.
Given parameters, it generates total mortality values from a log normal random walk
}
\details{
Random walk errors are first generated for years and ages independently.
The outer product of these errors are then calculated to generate a total
mortality matrix as follows:
\deqn{log(Z_{a, y}) = log(z_{a}) + \delta_{a, y}}{Z_a,y = z_a,y + e_a,y}
}
\examples{
simAbundance(Z = simZ(mean = 0.6, sd = 0.3))
simAbundance(Z = simZ(mean = list(ages = c(0.5, 0.3, 0.2)),
                      breaks = list(ages = c(1, 2)), sd = 0.3))
simAbundance(Z = simZ(mean = list(ages = c(0.5, 0.3, 0.2), years = c(0.3, 2, 0.3)),
                      breaks = list(ages = c(1, 2), years = c(4, 6)), sd = 0.3))

}

