% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{strat_data}
\alias{strat_data}
\title{Prepare simulated data for stratified analysis}
\usage{
strat_data(sim, length_group = 3, alk_scale = "division")
}
\arguments{
\item{sim}{Simulation object returned by \code{\link[=sim_survey]{sim_survey()}}.}

\item{length_group}{Size of the length frequency bins.}

\item{alk_scale}{Spatial scale at which to construct and apply age-length keys:
\code{"division"}, \code{"strat"}, or \code{"set"}.}
}
\value{
A list containing:
\itemize{
\item \code{setdet}: Set details
\item \code{lf}: Length-frequency data
\item \code{af}: Age-frequency data
}
}
\description{
Generate set details (\code{setdet}), length-frequency (\code{lf}), and age-frequency (\code{af})
data for stratified analysis.
}
