% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_grid}
\alias{survey_grid}
\title{Sample survey simulation grid}
\format{
A \code{stars} object with 4 attributes:
\itemize{
\item \strong{cell}: Survey cell identifier
\item \strong{division}: NAFO division
\item \strong{strat}: Survey strata number
\item \strong{depth}: Mean depth of the waters under each cell (in meters)
}

For more details on how this file was created, see the \code{data-raw} folder in this package.
}
\usage{
survey_grid
}
\description{
An exemplar of the structure of a survey grid object used by functions in this package.
}
\keyword{datasets}
