% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_surveys.R
\name{test_surveys}
\alias{test_surveys}
\title{Test sampling design of multiple surveys}
\usage{
test_surveys(sim, n_sims = 10, n_loops = 10, cores = 2, set_den = c(0.3,
  0.5, 0.8, 1, 2, 3, 6, 9)/1000, lengths_cap = c(2, 3, 5, 8, 10, 20, 30, 60,
  90, 100, 200, 400, 600, 1000), ages_cap = c(2, 3, 5, 8, 10, 20, 30, 60),
  ...)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}.}

\item{n_sims}{Number of times to simulate a survey over the simulated population.
Requesting a large number of simulations here may max out your RAM.}

\item{n_loops}{Number of times to run the \code{\link{sim_survey}} function. Total
simulations run will be the product of \code{n_sims} and \code{n_loops}
arguments. Low numbers of \code{n_sims} and high numbers of \code{n_loops}
will be easier on RAM, but slower.}

\item{cores}{Number of cores to use in parallel}

\item{set_den}{Vector of set densities (number of sets per [grid unit] squared)}

\item{lengths_cap}{Vector of maximum number of lengths measured per set}

\item{ages_cap}{Vector of maximum number of otoliths to collect per length group
per division per year}

\item{...}{Additional arguments to pass to \code{\link{sim_survey}}.}
}
\value{
Adds a table of survey designs tested. Also adds details and summary
        stats of stratified estimate error to the \code{sim} list, ending with
        \code{"_strat_error"} or \code{"_strat_error_stats"}. Error statistics
        includes mean absolute error (\code{"MAE"}), mean squared error (\code{"MSE"}),
        and root mean squared error (\code{"RMSE"}). Survey and stratified analysis
        details are not kept to minimize object size.
}
\description{
This function allows a series of sampling design settings to
be set and tested on the simulated population. All combinations of the
supplied settings (\code{set_den}, \code{lengths_cap}, \code{ages_cap}) are tested.
}
