% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{convert_N}
\alias{convert_N}
\title{Convert abundance-at-age matrix to abundance-at-length}
\usage{
convert_N(N_at_age = NULL, lak = NULL)
}
\arguments{
\item{N_at_age}{A matrix of abundance-at-age values.}

\item{lak}{A length-age key matrix — i.e., the probability of being in a specific
length group given age.}
}
\value{
A matrix of abundance-at-length values.
}
\description{
Converts an abundance-at-age matrix to an abundance-at-length matrix using
a length-age key. Both input matrices must be named appropriately.
}
