% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{error_stats}
\alias{error_stats}
\title{Calculate common error statistics}
\usage{
error_stats(error)
}
\arguments{
\item{error}{Vector of errors}
}
\value{
Returns a named vector of error statistics including mean absolute
        root mean squared error (\code{"RMSE"})
}
\description{
Calculate common error statistics
}
