% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_surveys.R
\name{expand_surveys}
\alias{expand_surveys}
\title{Set-up a series of surveys from all combinations of settings supplied}
\usage{
expand_surveys(set_den = c(0.3, 0.5, 0.8, 1, 2, 3, 6, 9)/1000,
  lengths_cap = c(2, 3, 5, 8, 10, 20, 30, 60, 90, 100, 200, 400, 600, 1000),
  ages_cap = c(2, 3, 5, 8, 10, 20, 30, 60))
}
\arguments{
\item{set_den}{Vector of set densities (number of sets per [grid unit] squared)}

\item{lengths_cap}{Vector of maximum number of lengths measured per set}

\item{ages_cap}{Vector of maximum number of otoliths to collect per length group
per division per year}
}
\description{
Function is simply a wrapper for \code{\link{base::expand.grid}} that
adds a survey number to the returned object
}
