% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fibonacci}
\alias{fibonacci}
\title{Generate Fibonacci sequence}
\usage{
fibonacci(from, to)
}
\arguments{
\item{from, to}{Approximate start and end values of the sequence}
}
\description{
Generate Fibonacci sequence
}
\examples{

fibonacci(2, 200)

}
