% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{group_lengths}
\alias{group_lengths}
\title{Convert length to length group}
\usage{
group_lengths(length, group)
}
\arguments{
\item{length}{Interval from \code{\link{base::findInterval}}}

\item{group}{Length group used to cut the length data}
}
\description{
Helper function for converting lengths to length groups
(Note: this isn't a general function; the output midpoints defining the
groups alligns with DFO specific method/labeling)
}
