% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculate intraclass correlation}
\usage{
icc(x, group)
}
\arguments{
\item{x}{Response variable}

\item{group}{Group}
}
\description{
This is a simple function for calculating intraclass correlation. It uses
\code{\link{lme4::lmer}} to run the formula described here:
https://en.wikipedia.org/wiki/Intraclass_correlation
}
