% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Southern Newfoundland coastline}
\format{A SpatialPolygons object

Derived from global administrative boundaries data (http://gadm.org/) downloaded
using the \code{\link{raster::getData}} function. Details provided in the
data-raw folder for this package.}
\usage{
land
}
\description{
Southern Newfoundland coastline
}
\keyword{datasets}

