% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{object_size}
\alias{object_size}
\title{Print object size}
\usage{
object_size(x, units = "Mb")
}
\arguments{
\item{x}{An R object.}

\item{units}{The units to be used when printing the size.}
}
\value{
A character string with the object size followed by the unit.
}
\description{
A wrapper for \code{\link[utils:object.size]{utils::object.size()}} that prints in megabytes (Mb) by default.
}
