% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_trend}
\alias{plot_trend}
\alias{plot_surface}
\alias{plot_distribution}
\title{Simple plotting functions}
\usage{
plot_trend(sim, ...)

plot_surface(sim, mat = "N", ...)

plot_distribution(sim, ages = 1, years = 1, type = "contour",
  axes = TRUE, scale = "natural", ...)
}
\arguments{
\item{sim}{Object returned by \code{\link{sim_abundance}},
\code{\link{sim_distribution}}, etc.}

\item{...}{Additional arguments to pass to \code{\link{plotly::plot_ly}}.}

\item{mat}{Name of matrix in \code{sim} list to plot.}

\item{ages}{Subset data to one or more ages.}

\item{years}{Subset data to one or more years.}

\item{scale}{Plot response on "natural" or "log" scale?}
}
\description{
These functions are simple plotting helpers to get some quick
visuals of values produced by \code{\link{sim_abundance}},
\code{\link{sim_distribution}}, etc.
}
