% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_trend}
\alias{plot_trend}
\alias{plot_surface}
\alias{plot_distribution}
\alias{plot_distribution_slider}
\alias{plot_error_surface}
\alias{plot_error_cross_sections}
\alias{plot_true_vs_est}
\alias{plot_samp_dist}
\alias{plot_total_abundance}
\alias{plot_set_catch}
\alias{plot_true_vs_est_fan}
\title{Simple plotting functions}
\usage{
plot_trend(sim, col = viridis::viridis(1), ...)

plot_surface(sim, mat = "N", ...)

plot_distribution(sim, ages = 1, years = 1, type = "contour",
  axes = TRUE, scale = "natural", ...)

plot_distribution_slider(sim, ages = 1, years = 1)

plot_error_surface(sim)

plot_error_cross_sections(sim, col = viridis::viridis(1))

plot_true_vs_est(sim, which_survey = 1, max_sims = 50, facet_by = "age",
  cols = viridis::viridis(3))

plot_samp_dist(sim, which_year = 1, which_sim = 1,
  col = viridis::viridis(1))

plot_total_abundance(sim, which_year = 1, main = "")

plot_set_catch(sim, which_year = 1, which_sim = 1, main = "")

plot_true_vs_est_fan(sim, surveys = 1, frame = "year", quants = seq(90,
  10, by = -10))
}
\arguments{
\item{sim}{Object returned by \code{\link{sim_abundance}},
\code{\link{sim_distribution}}, etc.}

\item{col}{Plot color}

\item{...}{Additional arguments to pass to \code{\link{plotly::plot_ly}}.}

\item{mat}{Name of matrix in \code{sim} list to plot.}

\item{ages}{Subset data to one or more ages.}

\item{years}{Subset data to one or more years.}

\item{scale}{Plot response on "natural" or "log" scale?}

\item{which_survey}{Subset to specific survey}

\item{max_sims}{Maximum number of sims to plot}

\item{facet_by}{Facet plot by "age" or "year"?}

\item{cols}{Plot colors}

\item{which_year}{Subset to specific year}

\item{which_sim}{Subset to specific sim}

\item{main}{Plot title}

\item{frame}{Frame plots by "age" or "year"?}

\item{quants}{Quantile intervals to display on fan plot}

\item{survey}{Subset data to one or more surveys.}
}
\description{
These functions are simple plotting helpers to get some quick
visuals of values produced by \code{\link{sim_abundance}},
\code{\link{sim_distribution}}, etc.
}
