% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samp.R
\name{round_sim}
\alias{round_sim}
\title{Round simulated population}
\usage{
round_sim(pop)
}
\arguments{
\item{pop}{Simulation from \code{\link{sim_distribution}}}
}
\description{
Round simulated population
}
