% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{round_sim}
\alias{round_sim}
\title{Round simulated population}
\usage{
round_sim(sim)
}
\arguments{
\item{sim}{A simulation object returned by \code{\link[=sim_distribution]{sim_distribution()}}.}
}
\value{
A rounded version of the simulation object.
}
\description{
Rounds values in a simulation object, typically used as a helper function within \code{\link[=sim_survey]{sim_survey()}}.
}
