% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{round_sim}
\alias{round_sim}
\title{Round simulated population}
\usage{
round_sim(sim)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}}
}
\value{
Returns a rounded simulation object. Largely used as a helper in \code{\link{sim_survey}}.
}
\description{
Round simulated population
}
