% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{run_strat}
\alias{run_strat}
\title{Run stratified analysis on simulated data}
\usage{
run_strat(sim)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_survey}}}
}
\value{
Adds stratified analysis results for the total population (\code{"total_strat"})
        and the population aggregated by length group and age (\code{"length_strat"} and
        \code{"age_strat"}, respectively) to the \code{sim} list.
}
\description{
Run stratified analysis on simulated data
}
\examples{

sim <- sim_abundance(ages = 1:10, years = 1:5, R = sim_R(mean = 1e+7),
                     growth = sim_vonB(length_group = 1)) \%>\%
           sim_distribution(grid = make_grid(res = c(10, 10)),
                            ays_covar = sim_ays_covar(sd = 1)) \%>\%
           sim_survey(n_sims = 10, q = sim_logistic(k = 2, x0 = 3)) \%>\%
           run_strat()

}
