% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_abundance}
\alias{sim_abundance}
\title{Simulate basic population dynamics model}
\usage{
sim_abundance(ages = 1:14, years = 1:20, Z = sim_Z(), R = sim_R(),
  N0 = sim_N0())
}
\arguments{
\item{ages}{Ages to include in the simulation.}

\item{years}{Years to include in the simulation.}

\item{Z}{Total mortality function, like \code{\link{sim_Z}}, for generating
mortality matrix.}

\item{R}{Recruitment (i.e. abundance at \code{min(ages)}) function, like
\code{\link{sim_R}}, for generating recruitment vector.}

\item{N0}{Starting abundance (i.e. abundance at \code{min(years)}) function, like
\code{\link{sim_N0}}, for generating starting abundance vector.}
}
\value{
A \code{list} of length 3:
\itemize{
  \item{\code{R} - Vector of recruitment values}
  \item{\code{N0} - Vector of starting abundance values}
  \item{\code{Z} - Matrix of total mortality values}
  \item{\code{N} - Matrix of abundance values}
}
}
\description{
Simulate basic population dynamics model
}
\details{
Abundance from is calculated using a standard population dynamics model:
\deqn{N_{a, y} = N_{a - 1, y - 1} * exp(-Z_{a - 1, y - 1})}{N_a,y = N_a-1,y-1 * exp(-Z_a-1,y-1)}
}
\examples{
sim_abundance()

}
