% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_distribution}
\alias{sim_distribution}
\title{Simulate spatial and temporal distribution}
\usage{
sim_distribution(pop = sim_abundance(), grid = sim_grid(),
  space_covar = sim_sp_covar(), ay_covar = sim_ay_covar(phi_age = 0.05,
  phi_year = 0), depth_par = parabolic_fun())
}
\arguments{
\item{pop}{An abundance at age matrix with ages defining the rows and years defining
the columns (i.e. same structure as a matrix provided by \code{\link{sim_abundance}})}

\item{grid}{A \code{\link{SpatialPolygonsDataFrame}} defining a regular or irregular
grid with the same structure as \code{\link{survey_grid}}}
}
\description{
Provided an abundance at age matrix (like one provided by \code{\link{sim_abundance}})
and a survey grid (like \code{\link{survey_grid}}) to populate, this function
applies correlated space, time and size error to simulate the distribution
of the population.
}
\examples{
sim_distribution()

}
