% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_distribution}
\alias{sim_distribution}
\title{Simulate spatial and temporal distribution}
\usage{
sim_distribution(pop = sim_abundance(), grid = survey_grid,
  space_covar = sim_covar(range = 200, variance = 0.5, model = "matern"),
  age_covar = "ran", year_covar = "rw", age_break = Inf,
  year_break = Inf, depth_par = parabolic_fun(mu = 200, sigma = 100),
  scale_error = TRUE)
}
\arguments{
\item{pop}{An abundance at age matrix with ages defining the rows and years defining
the columns (i.e. same structure as a matrix provided by \code{\link{sim_abundance}})}

\item{grid}{A \code{\link{SpatialPolygonsDataFrame}} defining a regular or irregular
grid with the same structure as \code{\link{survey_grid}}}
}
\description{
Provided an abundance at age matrix (like one provided by \code{\link{sim_abundance}})
and a survey grid (like \code{\link{survey_grid}}) to populate, this function
applies correlated space, time and size error to simulate the distribution
of the population.
}
\examples{
sim_distribution()

}
