% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_distribution}
\alias{sim_distribution}
\title{Simulate spatial and temporal distribution}
\usage{
sim_distribution(sim, grid = sim_grid(), ays_covar = sim_ays_covar(),
  depth_par = sim_parabola())
}
\arguments{
\item{sim}{An abundance at age matrix like one produced by \code{\link{sim_abundance}}}

\item{grid}{A raster object defining the survey grid, like \code{\link{survey_grid}}
or one produced by \code{\link{sim_grid}}}

\item{ays_covar}{Closure for simulating age-year-space covariance}

\item{depth_par}{Closure for defining relationship between abundance and depth}
}
\description{
Provided an abundance at age matrix (like one provided by \code{\link{sim_abundance}})
and a survey grid (like \code{\link{survey_grid}}) to populate, this function
applies correlated space, age and year error to simulate the distribution
of the population. The ability to simulate distributions by length is yet to be implemented.
}
\details{
This function simulates the probability of simulated fish inhabiting
a cell as a function of a parabolic relationship with depth and space, age,
and year autocorrelated errors.
}
\examples{
sim_distribution()

}
