% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samp.R
\name{sim_index}
\alias{sim_index}
\title{Simulate population available to the survey}
\usage{
sim_index(pop, n_sims = 1, q_mod = sim_logistic(), binom_error = FALSE)
}
\arguments{
\item{pop}{Simulation from \code{\link{sim_distribution}}, rounded by \code{\link{sim_distribution}}}

\item{n_sims}{Number of simulations to produce}

\item{q_mod}{Closure, such as \code{\link{sim_logistic}}, for simulating catchability at age
(returned values must be between 0 and 1)}

\item{binom_error}{Impose binomial error?}
}
\description{
Simulate population available to the survey
}
