% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_logistic}
\alias{sim_logistic}
\title{Closure for simulating logistic curve}
\usage{
sim_logistic(k = 2, x0 = 3, plot = FALSE)
}
\arguments{
\item{k}{The steepness of the curve}

\item{x0}{The x-value of the sigmoid's midpoint}

\item{plot}{Plot relationship}
}
\description{
This closure is useful for simulating q inside the
\code{\link{sim_survey}} function
}
\examples{
logistic_fun <- sim_logistic(k = 2, x0 = 3, plot = TRUE)
logistic_fun(x = 1:10)

}
