% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_parabola}
\alias{sim_parabola}
\title{Define relationships with covariates}
\usage{
sim_parabola(alpha = 0, mu = 250, sigma = 50, plot = FALSE)
}
\arguments{
\item{alpha, mu, sigma}{Parameters that control the shape of the parabola}

\item{plot}{Produce a simple plot of the simulated values?}
}
\description{
Closure to be used in \code{\link{sim_distribution}}
}
\examples{

parabola_fun <- sim_parabola(alpha = 25, mu = 50, sigma = 5, plot = TRUE)
parabola_fun(x = 0:100)

}
