% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop.R
\name{sim_R}
\alias{sim_R}
\alias{sim_Z}
\title{Simulate random recruitment and natural mortality}
\usage{
sim_R(mean = 20000, sd = 4)

sim_Z(mean = 0.4, sd = 1.1, breaks = NULL)
}
\arguments{
\item{mean, sd}{Mean and standard deviation}

\item{breaks}{Provide breaks to age and/or year series to specify group specific
mean total mortality values. If specified, multiple mean values must be
supplied; one more than the number of breaks. Provide named lists (see
examples below).}
}
\description{
These functions return a function to use inside \code{\link{sim_abundance}}.
Given parameters, it generates R and Z values.
}
\details{
Both functions simply generate uncorrelated recruitment or mortality
values from a log normal distribution.
}
\examples{
sim_abundance(R = sim_R(mean = 100000, sd = 4))
sim_abundance(Z = sim_Z(mean = 0.6, sd = 1.5))
sim_abundance(Z = sim_Z(mean = list(ages = c(0.5, 0.3, 0.2)),
                      breaks = list(ages = c(1, 2))))
sim_abundance(Z = sim_Z(mean = list(ages = c(0.5, 0.3, 0.2), years = c(0.3, 2, 0.3)),
                        breaks = list(ages = c(1, 2), years = c(4, 6))))

}

