% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sets.R
\name{sim_sets}
\alias{sim_sets}
\title{Simulate survey sets}
\usage{
sim_sets(pop = sim_distribution(), min_sets = 1, set_den = 5/1000,
  resample_units = FALSE)
}
\arguments{
\item{pop}{Simulation object from \code{\link{sim_distribution}}}

\item{min_sets}{Minimum number of sets per strat}

\item{set_den}{Set density}

\item{resample_units}{Sllow resampling of sampling units (grid cells)?}
}
\description{
Simulate survey sets
}
