% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_sets}
\alias{sim_sets}
\title{Simulate survey sets}
\usage{
sim_sets(sim, n_sims = 1, trawl_dim = c(1.5, 0.02), min_sets = 1,
  set_den = 5/1000, resample_cells = FALSE)
}
\arguments{
\item{sim}{Simulation object from \code{\link{sim_distribution}}}

\item{n_sims}{Number of simulations to produce}

\item{trawl_dim}{Trawl width and distance (same units as grid)}

\item{min_sets}{Minimum number of sets per strat}

\item{set_den}{Set density}

\item{resample_cells}{Allow resampling of sampling units (grid cells)?}
}
\description{
Simulate survey sets
}
