% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_sp_covar}
\alias{sim_sp_covar}
\alias{sim_ay_covar}
\title{Simulate exponential or Matern covariance}
\usage{
sim_sp_covar(range = 100, sd = 0.1, model = "matern")

sim_ay_covar(sd = 10, phi_age = 0.8, phi_year = 0.2)
}
\arguments{
\item{range}{Decorrelation range}

\item{sd}{Variance}

\item{model}{String indicating either "exponential" or "matern" as the correlation function}

\item{phi_age}{Defines autocorrelation through ages. Can be one value or a vector of the same
length as ages}

\item{phi_year}{Defines autocorrelation through years. Can be one value or a vector of the same
length as years}
}
\description{
These functions return a function to use inside \code{\link{sim_distribution}}.
}
