% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_survey}
\alias{sim_survey}
\title{Simulate stratified-random survey}
\usage{
sim_survey(sim, n_sims = 10, q = sim_logistic(), growth = sim_vonB(),
  resample_cells = FALSE, binom_error = FALSE, max_lengths = 100,
  length_group = 3, max_ages = 10)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}}

\item{n_sims}{Number of surveys to simulate over the simulated population. Note: requesting
a large number of simulations may max out your RAM}

\item{q}{Closure, such as \code{\link{sim_logistic}}, for simulating catchability at age
(returned values must be between 0 and 1)}

\item{growth}{Closure, such as \code{\link{sim_vonB}}, for simulating length given age}

\item{resample_cells}{Allow resampling of sampling units (grid cells)?}

\item{binom_error}{Impose binomial error?}

\item{max_lengths}{Maximum number of lengths measured per set}

\item{length_group}{Length group for otolith collection}

\item{max_ages}{Maximum number of otoliths to collect per length group per division per year}
}
\value{
A list including rounded population simulation, set locations and details
and sampling details. Note that that N = "true" population, I = population available
to the survey, n = number caught by survey.
}
\description{
Simulate stratified-random survey
}
