% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_survey}
\alias{sim_survey}
\title{Simulate stratified-random survey}
\usage{
sim_survey(sim, n_sims = 1, q = sim_logistic(), trawl_dim = c(1.5, 0.02),
  resample_cells = FALSE, binom_error = TRUE, min_sets = 2,
  set_den = 3/1000, lengths_cap = 400, length_group = 1, ages_cap = 10,
  age_sampling = "stratified", light = TRUE)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}}

\item{n_sims}{Number of surveys to simulate over the simulated population. Note: requesting
a large number of simulations may max out your RAM}

\item{q}{Closure, such as \code{\link{sim_logistic}}, for simulating catchability at age
(returned values must be between 0 and 1)}

\item{trawl_dim}{Trawl width and distance (same units as grid)}

\item{resample_cells}{Allow resampling of sampling units (grid cells)? Setting to TRUE may introduce bias
because depletion is imposed at the cell level.}

\item{binom_error}{Impose binomial error? Setting to FALSE may introduce bias in stratified estimates
at older ages because of more frequent rounding to zero.}

\item{min_sets}{Minimum number of sets per strat}

\item{set_den}{Set density (number of sets per [grid unit] squared). WARNING:
may return an error if \code{set_den} is high and
\code{resample_cells = FALSE} because the number of sets allocated may
exceed the number of cells in a strata.}

\item{lengths_cap}{Maximum number of lengths measured per set}

\item{length_group}{Length group for otolith collection}

\item{ages_cap}{If \code{age_sampling = "length stratified"}, this cap represents the
maximum number of otoliths to collect per length group
per division per year. If \code{age_sampling = "random"}, it is the
maximum number of otoliths to collect from measured fish per set.}

\item{age_sampling}{Should age sampling be length "stratified" or "random"?}

\item{light}{Drop some objects from the output to keep object size low?}
}
\value{
A list including rounded population simulation, set locations and details
and sampling details. Note that that N = "true" population, I = population available
to the survey, n = number caught by survey.
}
\description{
Simulate stratified-random survey
}
\examples{

sim <- sim_abundance(ages = 1:10, years = 1:5, R = sim_R(mean = 1e+7)) \%>\%
           sim_distribution(grid = make_grid(res = c(10, 10)),
                            ays_covar = sim_ays_covar(sd = 1)) \%>\%
           sim_survey(n_sims = 10, q = sim_logistic(k = 2, x0 = 3))
plot_survey(sim, which_year = 2, which_sim = 1)

}
