% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_survey_parallel}
\alias{sim_survey_parallel}
\title{Simulate stratified random surveys using parallel computation}
\usage{
sim_survey_parallel(sim, n_sims = 1, n_loops = 100, cores = 6,
  quiet = FALSE, ...)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}}

\item{n_sims}{Number of times to simulate a survey over the simulated population.
Requesting a large number of simulations here may max out your RAM.}

\item{n_loops}{Number of times to run the \code{\link{sim_survey}} function. Total
simulations run will be the product of \code{n_sims} and \code{n_loops}
arguments. Low numbers of \code{n_sims} and high numbers of \code{n_loops}
will be easier on RAM, but may be slower.}

\item{cores}{Number of cores to use in parallel. More cores should speed up the process.}

\item{quiet}{Print message on what to expect for duration?}

\item{...}{Arguments passed on to \code{sim_survey}
\describe{
  \item{sim}{Simulation from \code{\link{sim_distribution}}}
  \item{n_sims}{Number of surveys to simulate over the simulated population. Note: requesting
a large number of simulations may max out your RAM}
  \item{q}{Closure, such as \code{\link{sim_logistic}}, for simulating catchability at age
(returned values must be between 0 and 1)}
  \item{trawl_dim}{Trawl width and distance (same units as grid)}
  \item{resample_cells}{Allow resampling of sampling units (grid cells)?}
  \item{binom_error}{Impose binomial error? Setting to FALSE may introduce bias in stratified estimates
at older ages because of more frequent rounding to zero.}
  \item{min_sets}{Minimum number of sets per strat}
  \item{set_den}{Set density (number of sets per [grid unit] squared)}
  \item{lengths_cap}{Maximum number of lengths measured per set}
  \item{length_group}{Length group for otolith collection}
  \item{ages_cap}{Maximum number of otoliths to collect per length group
per division per year}
  \item{light}{Drop some objects from the output to keep object size low?}
}}
}
\description{
This function is a wrapper for \code{\link{sim_survey}} except it allows for
many more total itterations to be run than \code{\link{sim_survey}} before running
into RAM limitations.
}
\details{
\code{\link{sim_survey}} is hard-wired here to be "light" to minimize object size.
}
