% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survey.R
\name{sim_vonB}
\alias{sim_vonB}
\title{Closure for simulating length given age using von Bertalanffy notation}
\usage{
sim_vonB(Linf = 120, L0 = 5, K = 0.1, log_sd = 0.1, digits = 0,
  plot = FALSE)
}
\arguments{
\item{Linf}{Mean asymptotic length}

\item{L0}{Length at birth}

\item{K}{Growth rate parameter}

\item{log_sd}{Standard deviation of the relationship in log scale}

\item{digits}{Integer indicating the number of decimal places to round the values to}

\item{plot}{Produce a simple plot of the simulated values?}
}
\description{
Closure for simulating length given age using von Bertalanffy notation
}
