% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simAbundance}
\alias{simAbundance}
\title{Simulate basic population dynamics model}
\usage{
simAbundance(ages = 1:6, years = 1:10, Z = simZ(), R = simR())
}
\arguments{
\item{ages}{Ages to include in the simulation.}

\item{years}{Years to include in the simulation.}

\item{Z}{Total mortality function, like \code{\link{simZ}}, for generating
mortality matrix.}

\item{R}{Recruitment (i.e. Abundance at \code{min(ages)}) function, like
\code{\link{simR}}, for generating recruitment vector.}
}
\value{
A \code{list} of length 3:
\itemize{
  \item{\code{R} - Vector of recruitment values}
  \item{\code{Z} - Matrix of total mortality values}
  \item{\code{N} - Matrix of abundance values}
}
}
\description{
Simulate basic population dynamics model
}
\details{
Abundance from \code{ages[2:max(ages)]} and \code{years[2:max(years)]} is
calculated using a standard population dynamics model:
\deqn{N_{a, y} = N_{a - 1, y - 1} * exp(-Z_{a - 1, y - 1})}{N_a,y = N_a-1,y-1 * exp(-Z_a-1,y-1)}
Abundance at \code{min(ages)} is supplied by \code{r} and abundance at \code{ages[2:max(ages)]} are
calculated using the same equation above using \code{Z} values from \code{min(years)}.
}
\examples{
simAbundance()

}

