% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simR}
\alias{simR}
\title{Simulate random recruitment}
\usage{
simR(mean = 20000, sd = 4)
}
\arguments{
\item{mean, sd}{Mean and standard deviation of recruitment}
}
\description{
This function returns a function to use inside \code{\link{simAbundance}}.
Given parameters, it generates recruitment values from a log normal distribution
}
\examples{
simAbundance(R = simR(mean = 100000, sd = 4))

}

