% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPop.R
\name{simZ}
\alias{simZ}
\title{Simulate total mortality using a random walk}
\usage{
simZ(mean = 0.4, sd = 0.2)
}
\arguments{
\item{mean, sd}{Mean and standard deviation of total mortality}
}
\description{
This function returns a function to use inside \code{\link{simAbundance}}.
Given parameters, it generates total mortality values from a log normal random walk
}
\details{
Random walk errors are first generated for years and ages independently.
The outer product of these errors are then calculated to generate a total
mortality matrix.
}
\examples{
simAbundance(Z = simZ(mean = 0.6, sd = 0.3))

}

