% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{strat_data}
\alias{strat_data}
\title{Prepare simulated data for stratified analysis}
\usage{
strat_data(sim, length_group = 3, alk_scale = "division")
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_survey}}}

\item{length_group}{Size of the length frequency bins}

\item{alk_scale}{Spatial scale at which to construct and apply age-length-keys:
"division", "strat" or "set".}
}
\description{
Generate set details (setdet), length-frequency (lf)
and age-frequency (af) data for stratified analysis
}
