% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{strat_error}
\alias{strat_error}
\title{Calculate error of stratified estimates}
\usage{
strat_error(sim)
}
\arguments{
\item{sim}{Object from \code{\link{run_strat}} (includes simulated population and
survey along with stratified analysis results)}
}
\value{
Adds details and summary stats to \code{sim} list. The \code{"details"}
        object is a data.table that includes index estimates from the stratified
        analysis (\code{"I_hat"}) along with the simulated index (\code{"I"}). The
        \code{"summary"} object is a named vector including estimates of mean
        absolute error (\code{"MAE"}), mean squared error (\code{"MSE"}), and
        root mean squared error (\code{"RMSE"}).
}
\description{
Calculate error of stratified estimates
}
