% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_strat.R
\name{strat_means}
\alias{strat_means}
\title{Calculate stratified means, variances and confidence intervals across groups}
\usage{
strat_means(
  data = NULL,
  metric = NULL,
  strat_groups = NULL,
  survey_groups = NULL,
  confidence = 95
)
}
\arguments{
\item{data}{Expects data.table with all grouping variables in stacked format (must include
strat_area and tow_area for scaling values)}

\item{metric}{Variable in specificed data.table. e.g. "number", "mass"}

\item{strat_groups}{Grouping variables for calculations of the fine-scale strat-level
means (strat and strat_area are required). e.g. c("year", "species",
"shiptrip", "NAFOdiv", "strat", "strat_area","age")}

\item{survey_groups}{Grouping variables for large-scale summary calculations. e.g. ("year","species")}

\item{confidence}{Percent for confidence limits}
}
\description{
Calculate stratified means, variances and confidence intervals across groups
}
\details{
Function was mainly created for use in the \code{\link{run_strat}} function.
It first calculates strat-level statistics and then the larger-scale statistics like total abundance
}
