% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_grid}
\alias{survey_grid}
\title{Sample survey simulation grid.}
\format{A SpatialPolygonsDataFrame with 8 variables:
\describe{
  \item{cell}{Survey cell identifier}
  \item{division}{NAFO division}
  \item{strat}{Survey strata number}
  \item{easting}{Easting coordinates from UTM Zone 21, datum WGS84}
  \item{northing}{Northing coordinates from UTM Zone 21, datum WGS84}
  \item{area}{Area of Voronoi cell (polygon) the sampling unit lies within}
  \item{depth}{Mean depth of the waters under each Voronoi cell (polygon)}
}

For further details on how this file was created, see the data-raw folder for
this package}
\usage{
survey_grid
}
\description{
A exemplar for the structure of a survey grid object to supply to the functions
in this package.
}
\keyword{datasets}

