% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_grid}
\alias{survey_grid}
\title{Sample survey simulation grid.}
\format{A SpatialPolygonsDataFrame with 8 variables:
\describe{
  \item{id}{survey unit identifier}
  \item{division}{NAFO division}
  \item{strat}{survey strata number}
  \item{unit_num}{survey sampling unit number}
  \item{easting}{easting coordinates from UTM Zone 21}
  \item{northing}{northing coordinates from UTM Zone 21}
  \item{area}{area of Voronoi polygon the sampling unit lies within}
  \item{depth}{mean depth of the waters under each Voronoi polygon}
}

For further details on how this file was created, see the data-raw folder for
this package}
\usage{
survey_grid
}
\description{
A exemplar for the structure of a survey grid object to supply to the functions
in this package.
}
\keyword{datasets}

