% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_surveys.R
\name{test_surveys}
\alias{test_surveys}
\title{Test sampling design of multiple surveys}
\usage{
test_surveys(sim, surveys = expand_surveys(), keep_details = 1,
  n_sims = 1, n_loops = 100, cores = 2, export = NULL, ...)
}
\arguments{
\item{sim}{Simulation from \code{\link{sim_distribution}}.}

\item{surveys}{A data.frame or data.table with a sequence of surveys and their settings
with a format like the data.table returned by \code{\link{expand_surveys}}.}

\item{keep_details}{Keep details of a specific survey in the data.frame supplied to \code{surveys}.
Survey and stratified analysis details are dropped for the other surveys to
minimize object size.}

\item{n_sims}{Number of times to simulate a survey over the simulated population.
Requesting a large number of simulations here may max out your RAM.}

\item{n_loops}{Number of times to run the \code{\link{sim_survey}} function. Total
simulations run will be the product of \code{n_sims} and \code{n_loops}
arguments. Low numbers of \code{n_sims} and high numbers of \code{n_loops}
will be easier on RAM, but may be slower.}

\item{cores}{Number of cores to use in parallel. More cores should speed up the process.}

\item{export}{Directory for exporting results as they are generated. If NULL, nothing is
exported.}

\item{...}{Additional arguments to pass to \code{\link{sim_survey}}.}
}
\value{
Adds a table of survey designs tested. Also adds details and summary
        stats of stratified estimate error to the \code{sim} list, ending with
        \code{"_strat_error"} or \code{"_strat_error_stats"}. Error statistics
        includes mean absolute error (\code{"MAE"}), mean squared error (\code{"MSE"}),
        and root mean squared error (\code{"RMSE"}). Survey and stratified analysis
        details are not kept to minimize object size.
}
\description{
This function allows a series of sampling design settings to
be set and tested on the simulated population. All combinations of the
supplied settings (\code{set_den}, \code{lengths_cap}, \code{ages_cap}) are tested.
}
