% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_sim.R
\name{vis_sim}
\alias{vis_sim}
\title{Make a flexdashboard for visualizing the simulation}
\usage{
vis_sim(sim, ...)
}
\arguments{
\item{sim}{Object produced by \code{\link{sim_abundance}}, \code{\link{sim_distribution}},
\code{\link{sim_survey}} or \code{\link{test_surveys}}.}

\item{...}{Additional arguments to send to \link[rmarkdown]{run}}
}
\description{
Assumes the working directory is the project directory
}
